HLSManagerBase
